фиксация от 7.1.2022

+++ Digital_Pet_Passport/Digital_Pet_Passport/Model/Image.cs
@@ -1,4 +1,5 @@
 ﻿using Digital_Pet_Passport.Intefaces;
+using System.ComponentModel.DataAnnotations.Schema;
 using System.IO;
 namespace Digital_Pet_Passport.Model
 {
@@ -12,9 +13,22 @@ namespace Digital_Pet_Passport.Model
 
         public Pet Pet { get; set; }
 
+        [NotMapped]
+        public string NameGallerey { get; set; }
+
+        public const string ImageGallereyPets = "ImageGallereyPets";
+
+        public Image()
+        {
+            NameGallerey = $"{ImageGallereyPets} № {Id}";
+        }
+
         /// <summary>
         /// Сохранение изображения в приложении
         /// </summary>
+        /// <param name="path">Путь к файлу</param>
+        /// <param name="name">Имя нового файла</param>
+        /// <returns>Возвращает путь скопированного файла</returns>
         public string SaveImg(string path, string name)
         {
             string inst = new Logic.BaseWorkFiles().InRootPath;
@@ -26,6 +40,11 @@ namespace Digital_Pet_Passport.Model
                 ex.ToUpper() == ".jpeg".ToUpper())
             {
                 fullpath = System.IO.Path.Combine(inst, name.Trim() + ex);
+                if (File.Exists(fullpath))
+                {
+                    RemoveImage(fullpath);
+                 
+                }
                 fileInfo.CopyTo(fullpath);
                 return fullpath;
             }
@@ -33,19 +52,37 @@ namespace Digital_Pet_Passport.Model
             return "";
         }
 
+        /// <summary>
+        /// Копирует изображение во внутренние хранилище по указанному. Под свойства NameGallerey.
+        /// </summary>
+        /// <param name="path">Путь к файлу</param>
+        /// <returns>Возвращает путь скопированного файла</returns>
+        public string SaveImg(string path)
+        {
+            return SaveImg(path, NameGallerey);
+        }
+
+        /// <summary>
+        /// Удаление изображения по переданному пути
+        /// </summary>
+        /// <param name="path">Путь к файлу</param>
         public void RemoveImage(string path)
         {
-            if (File.Exists(path))
+            if (path != "DefoultPetImage.png")
             {
-                File.Delete(path);
-            }
+                if (File.Exists(path))
+                {
+                    File.Delete(path);
+                }
 
 #if DEBUG
-            if (File.Exists(path))
-            {
+                if (File.Exists(path))
+                {
 
-            }
+                }
 #endif
+            }
+
 
 
         }
